Duplicate the fd even if it's the same in the child
This is to fix an issue where we sometimes can pass an fd to our child
and sometimes not. If the source fd in the parent happens to be the same
as the fd what it would be in the parent, we previously did not dupe it.
Due to FD_CLOEXEC, this means the child would not be able to use the fd.

This could happen if your program opened a single fd you wanted to pass
to the child. This fd would then likely have fd 3. Then consider if you
want to pass it as fd 3 to the child. e.g., start(\@cmd, \undef,
$stdout, $stderr, $yourfd).

Instead, we now duplicate it so it will be usable in the child.