child fd inadvertently closed if TFDs need to be moved
If the parent process has an open fd which is the first in a sequence
of fds that will be used by the child process, the child fd accounting
code loses track of which fds to close, and ends up closing one needed
by the eventual sub-process.

Assume that the parent has fd=3 open. The child requires fd=3 & fd=4.
The parent opens fds 4 and 5 to handle the child's fds, as
it can't perform a direct map because of fd=3 being open.

The parent creates OPS entries with

OPS[0]: { TFD => 4, KFD => 3 }
OPS[1]: { TFD => 5, KFD => 4 }

In the child, fd=3 is closed as it is not needed.

The child handles the moving of fds in _do_kid_and_exit:

   2581   ## Lazy closing is so the same fd (ie the same TFD value) can be dup2'ed on
   2582   ## several times.
   2583   my @lazy_close;
   2584   for ( @{ $kid->{OPS} } ) {
   2585       if ( defined $_->{TFD} ) {
   2586           unless ( $_->{TFD} == $_->{KFD} ) {
   2587               $self->_dup2_gently( $kid->{OPS}, $_->{TFD}, $_->{KFD} );
   2588               push @lazy_close, $_->{TFD};
   2589           }
   2590       }
   ....
   2607   }

When servicing OPS[0], _dup2_gently dups fd=4 to fd=3 and line 2604
records fd=4 as needing to be closed.

When servicing OPS[1], _dup2_gently notices that OPS[0]{TFD} conflicts
with OPS[1]{KFD} and dups OPS[0]{TFD} elsewhere. It then dups fd=5
to fd=4 and line 2604 records fd=5 as needing to be closed.

At this point the fds marked in @lazy_close are closed:

   2609   for (@lazy_close) {
   2610       unless ( $closed[$_] ) {
   2611           _close($_);
   2612           $closed[$_] = 1;
   2613       }
   2614   }

and fd=4 gets closed even though it is needed by OPS[1].

So that's one bug. The other bug is that @lazy_close doesn't know
that OPS[0]{TFD} got moved (that was done in _dup2_gently), so that
fd does not get closed.

There's a quick and dirty work-around to fix the first problem, but I
think there is a more substantial problem at work here. There are
multiple different places that information about fds are kept:

* @fds
* @needed
* @closed
* @lazy_close

This can be simplified. The following notes are for the child process
only; I haven't analyzed the code for the parent process. This commit
doesn't affect the parent process.

* @fds is the most important data structure; it records the known open
  fds. It is already consistently managed by the _close, _dup, and
  _dup2_rudely routines.

* @needed is useful, but because it is separate from @fds, it can get
  out of sync. it is also local to a single function
  (_do_kid_and_exit) when it needs to be accessible elsewhere
  (e.g. _dup2_gently)

* There's no need for @closed; if the fd is not in @fds, it's not open.

* @lazy_close means that an fd is not needed, but don't delete it
  right away. There are two aspects to this. First, there's already
  the @needed array, so this duplicates that functionality.  Second,
  @lazy_close is never used as a guard to *prevent* closing an fd. So
  it really has no use. Simply recording the fd in @needed as not
  being needed is sufficient.

This commit does a couple of things.

* It uses @fds as the *only* repository of information about fds, so
  all information is centralized. @fds is now an array of hashes.

* @needed is replaced by the {needed} attribute in an @fds element,
  e.g. fds[0]{needed} = 1.

* Since @fds is the canonical list of fds and is managed correctly,
  @closed is not needed.

* @lazy_close is also no longer needed; the {needed} attribute for an
  fd is set to 0 so that it will be closed during the next sweep.

* the code is more explicit about specifying which fds are needed or
  not. In particular, it explicitly marks KFDs as needed after they
  are created. This prevents inadvertent closing of KFD's whose fds
  were previously TFDs. When a TFD fd is moved by _dup2_gently the new
  entry in @fds defaults to {needed} = 0, and it will get cleaned up
  during the next fd closing sweep.