Detect child exit faster on Windows, via lower select() timeout.
Windows lacks SIGCHLD.  When _select_loop() called select() with zero
file descriptors, IPC::Run would not respond to child termination until
the select() timeout elapsed.  For that case, use the 0.01s timeout
every time, as an exception to the usual doubling of the timeout until
we reach 0.5s.  This spends CPU to achieve responsiveness.